decoder: CF1 stream support 34bit - 32bit compile [1/1]
PD#SWPL-177304

Problem:
decoder support stream 34bit

Solution:
config stream prefix

Verify:
s6_s905x5_bl201

+++ drivers/frame_provider/decoder/avs_multi/avs_multi.c
@@ -2818,8 +2818,8 @@ static void check_timer_func(struct timer_list *timer)
 				READ_VREG(VLD_MEM_VIFIFO_LEVEL),
 				READ_VREG(VLD_MEM_VIFIFO_WP),
 				READ_VREG(VLD_MEM_VIFIFO_RP),
-				STBUF_READ(&vdec->vbuf, get_rp),
-				STBUF_READ(&vdec->vbuf, get_wp));
+				(u64)STBUF_READ(&vdec->vbuf, get_rp),
+				(u64)STBUF_READ(&vdec->vbuf, get_wp));
 		}
 	}
